Add download/restore functionality and UI updates
This commit adds a new "manage data" page where it is possible to
download and restore all data.

Also included are a number of UI improvements such as font-awesome icons
and a modal that explains what "save data in the browser" means.

The transaction upload has been nested under the transactions page.