More small tweaks (#50)
* Add better number parsing

This commit adds slightly better number parsing. It would be best with
i18n parsing, but we cannot always rely on browsers to supply the right
locale. This commit implements a simplistic approach to number parsing
which will currently work for the target languages :-)

* Improve classification/persistence, add migration

This commit improves the app in three areas:

1. A migration system is set up to support migration of data between
   versions. The first included migration adds a cleaned description.
2. Classification has been improved by using the cleaned description for
   training the naive bayes classifier.
3. Some problems with local storage persistence when restoring data has
   been adressed. Still not perfect though.

Closes #45