Add currency fetch and setting default currency
This commit adds functionality for fetching the Euro foreign exchange
rates as well as a dropdown for selecting the default currency based on
the supported currencies.

The exchange rates are downloaded and unzipped in a lambda function to
act as a proxy for the user for privacy and to avoid dealing with
zip-files on the client-side.

Other changes:
- Pluralize urls and names for charts and transactions.